Move mockupsmakers package to internal package
