+++ app/src/main/java/com/shapeapp/shape/internal/mockupsmakers/CardMockups.kt
@@ -1,4 +1,4 @@
-package com.shapeapp.shape.mockupsmakers
+package com.shapeapp.shape.internal.mockupsmakers
 
 import com.shapeapp.shape.data.entities.Card
 import java.util.*
