+++ app/src/main/java/com/shapeapp/shape/internal/mockupsmakers/DrawablesMockups.kt
@@ -1,4 +1,4 @@
-package com.shapeapp.shape.mockupsmakers
+package com.shapeapp.shape.internal.mockupsmakers
 
 import java.util.*
